pull logging functions out of the tskmgr codes and import this code
