+++ ena_build/dask_tskmgr.py
@@ -11,33 +11,9 @@ import dask
 from distributed import Client, as_completed
 
 import mysql_database
+from workflow_logging import setup_logger, clean_logger
 from dask_tasks import process_many_files
 from glob_tasks import glob_subdirs, glob_files
-
-###############################################################################
-# Logging Functions
-###############################################################################
-
-def setup_logger(name, log_file, level=logging.INFO):
-    """To setup as many loggers as you want"""
-    formatter = logging.Formatter('%(asctime)s    %(levelname)s       %(message)s')
-    handler = logging.FileHandler(log_file)
-    handler.setFormatter(formatter)
-
-    logger = logging.getLogger(name)
-    logger.setLevel(level)
-    logger.addHandler(handler)
-
-    return logger
-
-
-def clean_logger(logger):
-    """To cleanup the logger instances once we are done with them"""
-    for handle in logger.handlers:
-        handle.flush()
-        handle.close()
-        logger.removeHandler(handle)
-
 
 ###############################################################################
 # Parse Input Arguments and Files
