+++ ena_build/workflow_logging.py
@@ -1,0 +1,28 @@
+
+import logging
+
+###############################################################################
+# Logging Functions
+###############################################################################
+
+def setup_logger(name, log_file, level=logging.INFO):
+    """To setup as many loggers as you want"""
+    formatter = logging.Formatter('%(asctime)s    %(levelname)s       %(message)s')
+    handler = logging.FileHandler(log_file)
+    handler.setFormatter(formatter)
+
+    logger = logging.getLogger(name)
+    logger.setLevel(level)
+    logger.addHandler(handler)
+
+    return logger
+
+
+def clean_logger(logger):
+    """To cleanup the logger instances once we are done with them"""
+    for handle in logger.handlers:
+        handle.flush()
+        handle.close()
+        logger.removeHandler(handle)
+
+
